Next code point.  Arrows are showing up.  Using the rabbit icon color magenta for Arrows.  Still using the bear icon for cougars but colored  yellow.
Problem right now, is that I'm always getting a smaller, non magenta  rabbit icon near the center of the radar screen.  I've come to the conclusion it is not likely related to the Arrows code.  It's smaller.  It's not magenta.  And I can drop / eliminate all of the Arrows in a sceen and it still appears.  I can turn off all animals and spraycan marker to "No" in the options screen and the rabbit still appears.